# openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Wharf Backend API",
        "description": "Zero-knowledge sync backend for Wharf. The server only ever stores\nciphertext vault blobs and bcrypt hashes of client-derived credential\nkeys. It never sees passwords or plaintext vault contents.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:18080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/projects/{id}/vault": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Fetch the project vault and the caller's wrapped DEK (any member)",
                "operationId": "getProjectVault",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Project not found or caller is not a member",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectVaultResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectVaultResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Vault returned",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectVaultResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Replace the project vault with optimistic concurrency (keyed member)",
                "operationId": "updateProjectVault",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVaultRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "403": {
                        "description": "Caller holds no wrapped key yet",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Vault updated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found or caller is not a member",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Version conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Vault blob too large",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List the caller's projects",
                "operationId": "listProjects",
                "responses": {
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectSummary"
                                    }
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Projects returned",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectSummary"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a project with its initial vault and the owner's wrapped DEK",
                "operationId": "createProject",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProjectRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Project created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "412": {
                        "description": "Caller has not published a public key",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid vault or wrapped key",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}/rotate": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Rotate the project DEK and re-wrap it for all remaining members (admin+)",
                "operationId": "rotateProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RotateProjectRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Rotated; returns the new vault version",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found or caller is not a member",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Caller is not an admin, or an admin removing an admin",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid removal or wrapped key",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Version conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/VaultUpdateResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}/members/{userId}/key": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Seal the project DEK to a member's public key (admin+)",
                "operationId": "submitMemberKey",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SubmitWrappedKeyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Wrapped key is not 80 bytes"
                    },
                    "409": {
                        "description": "Vault version has moved on (stale DEK)"
                    },
                    "403": {
                        "description": "Caller is not an admin"
                    },
                    "404": {
                        "description": "Project/caller/target member not found"
                    },
                    "401": {
                        "description": "Not authenticated"
                    },
                    "204": {
                        "description": "Key stored"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/projects/{id}/invites": {
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Invite an email to the project (admin+)",
                "operationId": "createInvite",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInviteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Invite created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectInvite"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Caller is not an admin",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectInvite"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Project not found or caller is not a member",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectInvite"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Not authenticated",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectInvite"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Email is already a member or already invited",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectInvite"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/projects/{id}/members/me": {
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Leave a project (the owner must transfer ownership first)",
                "operationId": "leaveProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Project not found or caller is not a member"
                    },
                    "401": {
                        "description": "Not authenticated"
                    },
                    "409": {
                        "description": "Owner cannot leave without transferring ownership"
                    },
                    "204": {
                        "description": "Left the project"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectVaultResponse": {
                "type": "object",
                "properties": {
                    "vault": {
                        "type": "string",
                        "description": "Base64-encoded ciphertext project vault blob"
                    },
                    "version": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "wrappedDek": {
                        "type": "string",
                        "description": "The caller's base64 wrapped DEK, or null if they are still awaiting a key"
                    }
                },
                "description": "The project's ciphertext vault plus the caller's wrapped DEK"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Identity (access) token issued by /api/v1/auth.",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}